ti-sgx-ddk-um: Install rc.pvr if its not installed
Newer ti-sgx-ddk-um does not install it into sysconfdir

+++ sources/meta-yoe/dynamic-layers/meta-ti/recipes-graphics/libgles/ti-sgx-ddk-um_1.17.4948957.bbappend
@@ -10,8 +10,12 @@ do_install:append() {
     rm -rf ${D}${libdir}/libGLESv2.so.1
     ln -sf libGLESv2.so.2 ${D}${libdir}/libGLESv2.so.1
     if [ ${@bb.utils.contains('DISTRO_FEATURES', 'systemd', 'true', 'false', d)} ]; then
-        mv ${D}${sysconfdir}/init.d/rc.pvr ${D}${bindir}
-        rmdir ${D}${sysconfdir}/init.d ${D}${sysconfdir}
+        if [-e ${D}${sysconfdir}/init.d/rc.pvr]; then
+            mv ${D}${sysconfdir}/init.d/rc.pvr ${D}${bindir}
+            rmdir ${D}${sysconfdir}/init.d ${D}${sysconfdir}
+        else
+            install -Dm0755 ${S}/targetfs/common/etc/init.d/rc.pvr ${D}${bindir}/rc.pvr
+        fi
         install -Dm 644 ${WORKDIR}/rc.pvr.service ${D}/${systemd_unitdir}/system/rc.pvr.service
     fi
 }
